refactor the healpix code into separate modules
